requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
groq
requests
exa-py
newspaper4k
lxml_html_clean
sqlalchemy 